Initial commit
Following same format of using SBA 318 as basis for 319, rescoped the API to a more reasonable goal, focusing on working routes and functionality.  Secondary focus on code readability, project organization, and good commenting practices as part of my ongoing efforts to be a better contributor by thwarting my natural racoon-in-trashcan style coding habits.