cleanup
ditching files that had a purpose in sba 318 but are now just clutter.  Building out routes, configuring attached atlas db to reliably handle the connection and reseed sample data on startup.  Hopefully the single view page I left will be able to correctly render the view itself in the palette in the request to offer an immediate preview for how it'd look on a basic webpage.  Also started documenting the routes and data types in readme.